plumbing all hooked up
