+++ ui/src/app/utils/workspace-permissions.ts
@@ -5,6 +5,10 @@ export namespace WorkspacePermissionsUtil {
     return accessLevel === WorkspaceAccessLevel.OWNER ||
       accessLevel === WorkspaceAccessLevel.WRITER;
   }
+
+  export function isOwner(accessLevel: WorkspaceAccessLevel) {
+    return accessLevel === WorkspaceAccessLevel.OWNER;
+  }
 }
 
 export class WorkspacePermissions {
@@ -15,11 +19,6 @@ export class WorkspacePermissions {
     this.workspace = workspaceResponse.workspace;
     this.accessLevel = workspaceResponse.accessLevel;
   }
-
-  get isOwner() {
-    return this.accessLevel === WorkspaceAccessLevel.OWNER;
-  }
-
   get canWrite() {
     return WorkspacePermissionsUtil.canWrite(this.accessLevel);
   }
@@ -32,15 +31,5 @@ export class WorkspacePermissions {
 
   get isReadOnly() {
     return this.accessLevel === WorkspaceAccessLevel.READER;
-  }
-
-  get isPending() {
-    return this.workspace.researchPurpose.reviewRequested === true &&
-      this.workspace.researchPurpose.approved === null;
-  }
-
-  get isRejected() {
-    return this.workspace.researchPurpose.reviewRequested === true &&
-      this.workspace.researchPurpose.approved === false;
   }
 }
